go/common/crypto: Bump ed25519 version
