+++ .changelog/3458.breaking.md
@@ -1,0 +1,1 @@
+go/common/crypto: Bump ed25519 version
