+++ go/common/crypto/signature/signature.go
@@ -697,27 +697,6 @@ func BuildPublicKeyBlacklist(allowTestKeys bool) {
 	}
 
 	// Explicitly forbid other keys here.
-
-	// Small order points.
-	// See: https://github.com/jedisct1/libsodium/blob/master/src/libsodium/crypto_core/ed25519/ref10/ed25519_ref10.c#L1019.
-	for _, v := range []string{
-		// 0 (order 4).
-		"0000000000000000000000000000000000000000000000000000000000000000",
-		// 1 (order 1).
-		"0100000000000000000000000000000000000000000000000000000000000000",
-		// 2707385501144840649318225287225658788936804267575313519463743609750303402022 (order 8).
-		"26e8958fc2b227b045c3f489f2ef98f0d5dfac05d3c63339b13802886d53fc05",
-		// 55188659117513257062467267217118295137698188065244968500265048394206261417927 (order 8).
-		"c7176a703d4dd84fba3c0b760d10670f2a2053fa2c39ccc64ec7fd7792ac037a",
-		// p-1 (order 2).
-		"ecffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff7f",
-		// p (=0, order 4).
-		"edffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff7f",
-		// p+1 (=1, order 1).
-		"eeffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff7f",
-	} {
-		NewBlacklistedPublicKey(v)
-	}
 }
 
 // NewBlacklistedPublicKey creates a new blacklisted public key from the given
